requirements.txt
# Django dependency
django==5.0.1
django-environ==0.11.2
gunicorn==21.2.0
pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.0

# Utilities
